+++ src/core/jsproxy.c
@@ -67,6 +67,7 @@ Js_IDENTIFIER(get);
 Js_IDENTIFIER(set);
 Js_IDENTIFIER(delete);
 Js_IDENTIFIER(includes);
+_Py_IDENTIFIER(fileno);
 
 static PyObject* asyncio_get_event_loop;
 static PyTypeObject* PyExc_BaseException_Type;
@@ -1435,7 +1436,7 @@ check_buffer_compatibility(JsProxy* self, Py_buffer view, bool safe, bool dir)
 /**
  * Assign from a js buffer to a py buffer
  * obj -- A JsBuffer (meaning a PyProxy of an ArrayBuffer or an ArrayBufferView)
- * buffer -- A PyObject whcih supports the buffer protocol and is writable.
+ * buffer -- A PyObject which supports the buffer protocol and is writable.
  */
 static PyObject*
 JsBuffer_assign_to(PyObject* obj, PyObject* target)
@@ -1643,6 +1644,73 @@ static PyMethodDef JsBuffer_tobytes_MethodDef = {
   METH_NOARGS,
 };
 
+static long
+get_fileno(PyObject* file)
+{
+  PyObject* pyfileno = _PyObject_CallMethodIdNoArgs(file, &PyId_fileno);
+  if (pyfileno == NULL) {
+    return -1;
+  }
+  return PyLong_AsLong(pyfileno);
+}
+
+static PyObject*
+JsBuffer_write_to_file(PyObject* jsbuffer, PyObject* file)
+{
+  int fd = get_fileno(file);
+  if (fd == -1) {
+    return NULL;
+  }
+  if (hiwire_write_to_file(JsProxy_REF(jsbuffer), fd)) {
+    return NULL;
+  }
+  Py_RETURN_NONE;
+}
+
+static PyMethodDef JsBuffer_write_to_file_MethodDef = {
+  "to_file",
+  (PyCFunction)JsBuffer_write_to_file,
+  METH_O,
+};
+
+static PyObject*
+JsBuffer_read_from_file(PyObject* jsbuffer, PyObject* file)
+{
+  int fd = get_fileno(file);
+  if (fd == -1) {
+    return NULL;
+  }
+  if (hiwire_read_from_file(JsProxy_REF(jsbuffer), fd)) {
+    return NULL;
+  }
+  Py_RETURN_NONE;
+}
+
+static PyMethodDef JsBuffer_read_from_file_MethodDef = {
+  "from_file",
+  (PyCFunction)JsBuffer_read_from_file,
+  METH_O,
+};
+
+static PyObject*
+JsBuffer_into_file(PyObject* jsbuffer, PyObject* file)
+{
+  int fd = get_fileno(file);
+  if (fd == -1) {
+    return NULL;
+  }
+  if (hiwire_into_file(JsProxy_REF(jsbuffer), fd)) {
+    return NULL;
+  }
+  Py_RETURN_NONE;
+}
+
+static PyMethodDef JsBuffer_into_file_MethodDef = {
+  "_into_file",
+  (PyCFunction)JsBuffer_into_file,
+  METH_O,
+};
+
 static PyObject*
 JsBuffer_tostring(PyObject* self,
                   PyObject* const* args,
@@ -1798,6 +1866,9 @@ JsProxy_create_subtype(int flags)
     methods[cur_method++] = JsBuffer_tomemoryview_MethodDef;
     methods[cur_method++] = JsBuffer_tobytes_MethodDef;
     methods[cur_method++] = JsBuffer_tostring_MethodDef;
+    methods[cur_method++] = JsBuffer_write_to_file_MethodDef;
+    methods[cur_method++] = JsBuffer_read_from_file_MethodDef;
+    methods[cur_method++] = JsBuffer_into_file_MethodDef;
   }
   methods[cur_method++] = (PyMethodDef){ 0 };
   members[cur_member++] = (PyMemberDef){ 0 };
@@ -2035,6 +2106,9 @@ JsProxy_init(PyObject* core_module)
   SET_DOCSTRING(JsBuffer_tomemoryview_MethodDef);
   SET_DOCSTRING(JsBuffer_tobytes_MethodDef);
   SET_DOCSTRING(JsBuffer_tostring_MethodDef);
+  SET_DOCSTRING(JsBuffer_write_to_file_MethodDef);
+  SET_DOCSTRING(JsBuffer_read_from_file_MethodDef);
+  SET_DOCSTRING(JsBuffer_into_file_MethodDef);
 #undef SET_DOCSTRING
 
   asyncio_module = PyImport_ImportModule("asyncio");