+++ src/py/_pyodide/_core_docs.py
@@ -1,4 +1,5 @@
 from typing import Any, Callable, Iterable
+from io import IOBase
 
 # All docstrings for public `core` APIs should be extracted from here. We use
 # the utilities in `docstring.py` and `docstring.c` to format them
@@ -122,6 +123,81 @@ class JsProxy:
         Copies the data once.
         Present only if the wrapped Javascript object is an ArrayBuffer or
         an ArrayBuffer view.
+        """
+
+    def to_file(self, file: IOBase):
+        """Writes the entire buffer to a file.
+
+        Will write the entire contents of the buffer to the current position of
+        the file.
+
+        Present only if the wrapped Javascript object is an ArrayBuffer or an
+        ArrayBuffer view.
+
+        Example
+        ------------
+        >>> import pytest; pytest.skip()
+        >>> from js import Uint8Array
+        >>> x = Uint8Array.new(range(10))
+        >>> with open('file.bin', 'wb') as fh:
+        ...    x.to_file(fh)
+        which is equivalent to,
+        >>> with open('file.bin', 'wb') as fh:
+        ...    data = x.to_bytes()
+        ...    fh.write(data)
+        but the latter copies the data twice whereas the former only copies the
+        data once.
+        """
+
+    def from_file(self, file: IOBase):
+        """Reads from a file into the buffer.
+
+        Will try to read a chunk of data the same size as the buffer from
+        the current position of the file.
+
+        Present only if the wrapped Javascript object is an ArrayBuffer or an
+        ArrayBuffer view.
+
+        Example
+        ------------
+        >>> import pytest; pytest.skip()
+        >>> from js import Uint8Array
+        >>> # the JsProxy need to be pre-allocated
+        >>> x = Uint8Array.new(range(10))
+        >>> with open('file.bin', 'rb') as fh:
+        ...    x.read_file(fh)
+        which is equivalent to
+        >>> x = Uint8Array.new(range(10))
+        >>> with open('file.bin', 'rb') as fh:
+        ...    chunk = fh.read(size=x.byteLength)
+        ...    x.assign(chunk)
+        but the latter copies the data twice whereas the former only copies the
+        data once.
+        """
+
+    def _into_file(self, file: IOBase):
+        """Will write the entire contents of the buffer to the current position
+        of the file using ``canOwn : true``. If the file is in the
+        memfs, the data does not need to be copied --  instead of copying the data,
+        the buffer will just be added to the backing for the file. After this, the
+        buffer cannot be used again.
+
+        Present only if the wrapped Javascript object is an ArrayBuffer or an
+        ArrayBuffer view.
+
+        Example
+        ------------
+        >>> import pytest; pytest.skip()
+        >>> from js import Uint8Array
+        >>> x = Uint8Array.new(range(10))
+        >>> with open('file.bin', 'wb') as fh:
+        ...    x._into_file(fh)
+        which is similar to
+        >>> with open('file.bin', 'wb') as fh:
+        ...    data = x.to_bytes()
+        ...    fh.write(data)
+        but the latter copies the data once whereas the former doesn't copy the
+        data.
         """
 
     def to_string(self, encoding=None) -> str: